requirements.txt
aiohttp==3.8.4
osm2geojson==0.2.3
PyYAML==6.0
shapely==2.0.1
sentry-sdk==1.25.1
